fastapi-server/requirements.txt
anyio==3.6.2
apturl==0.5.2
attrs==21.2.0
Automat==20.2.0
autopep8==2.0.2
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachelib==0.9.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.3
cloud-init==23.1.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fastapi==0.95.2
fasteners==0.14.1
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Cors==3.0.10
future==0.18.2
greenlet==2.0.2
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonify==0.5
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
pydantic==1.10.7
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-magic==0.4.24
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sos==4.4
SQLAlchemy==2.0.14
ssh-import-id==5.11
starlette==0.27.0
systemd-python==234
texttable==1.6.4
tomli==2.0.1
Twisted==22.1.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.22.0
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==2.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
